Fix PropertyService interface to use int instead of double

+++ src/main/java/com/example/rentHUB/service/PropertyService.java
@@ -6,12 +6,12 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 public interface PropertyService {
-    Property saveProperty(String title, double price, double oldPrice,
+    Property saveProperty(String title, int price, int oldPrice,
                           String deliveryInfo, MultipartFile image);
     Property savePropertyFromJson(
             String title,
-            double price,
-            double oldPrice,
+            int price,
+            int oldPrice,
             String deliveryInfo,
             String discount,
             String uploadedBy,
